+++ Back-End/src/main/java/com/ecoprint/control_center/repository/ResiduoRepository.java
@@ -3,7 +3,7 @@ package com.ecoprint.control_center.repository;
 import com.ecoprint.control_center.model.Residuo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ResiduoRepository extends JpaRepository<Residuo, Long>{
+public interface ResiduoRepository extends JpaRepository<Residuo, Integer>{
