+++ Back-End/src/main/java/com/ecoprint/control_center/repository/TipoResiduoRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
 
-public interface TipoResiduoRepository extends JpaRepository<TipoResiduo, Long>{
+public interface TipoResiduoRepository extends JpaRepository<TipoResiduo, Integer>{
 
     // Verifica se existe um tipo ou descrição já cadastrados
     @Query("SELECT COUNT(t) > 0 FROM tipo_residuo t WHERE t.tipo = :tipo OR t.descricao = :descricao")
